+++ src/render/python/emitter.cpp
@@ -12,5 +12,6 @@ MI_PY_EXPORT(EmitterExtras) {
         .def_value(EmitterFlags, SpatiallyVarying)
         .def_value(EmitterFlags, Invisible)
         .def_value(EmitterFlags, Portal)
+        .def_value(EmitterFlags, SamplesTexture)
         .def_value(EmitterFlags, Delta);
 }